Fix some indentation of code in Scala cheat sheet
The `}` belongs to the end of the `class`, not to the end of `def toString`

And the “abstract class” comment was arbitrarily at a different distance from the text from the “abstract method” comment. Either they should both be normal comments, with one space, or aligned. Since they were aligned weirdly, they were probably meant to be aligned, but the code earlier on the line was edited and that threw them off.